Added support for DEGIRO in App OTP
When no 2FA is provided, DEGIRO is now providing an 'In App' approval.

This PR provides support for this new approach. The `connection.py` example and the documentation have been also updated to reflect those changes

Support to request the OTP via SMS is still pending and may be part of a different PR